Implement typed partial function
Added an implementation of partial which can be used while remaining
fully type checked.

The new functions are separated by number of arguments.

Instead of using lambdas in a Pipe to control which arguments to bind,
now users can use a function like `partial3` which allows for continued
support of the type checker.